type casting stuff

+++ recursion/5-sqrt_recursion.c
@@ -10,7 +10,7 @@ int _sqrt_recursion(int n)
 {
 	if (n == 1)
 		return (1);
-	return (_sqrt_part_two(n, n / 2));
+	return ((int) (_sqrt_part_two((long int) n, (long int) n / 2)));
 }
 
 /**
